bug fixes and replace doubles quotes with string.Empty

+++ windows-client/DbUtils/MiscDBUtil.cs
@@ -309,7 +309,7 @@ namespace windows_client.DbUtils
         {
             if (msisdn == App.MSISDN)
                 msisdn = HikeConstants.MY_PROFILE_PIC;
-            
+
             if (imageBytes == null)
                 return;
 
@@ -324,7 +324,7 @@ namespace windows_client.DbUtils
                     {
                         if (isUpdated && store.FileExists(FileName + HikeConstants.FULL_VIEW_IMAGE_PREFIX))
                             store.DeleteFile(FileName + HikeConstants.FULL_VIEW_IMAGE_PREFIX);
-                        
+
                         using (FileStream stream = new IsolatedStorageFileStream(FileName, FileMode.Create, FileAccess.Write, FileShare.ReadWrite, store))
                         {
                             stream.Write(imageBytes, 0, imageBytes.Length);
@@ -344,7 +344,7 @@ namespace windows_client.DbUtils
         {
             if (msisdn == App.MSISDN)
                 msisdn = HikeConstants.MY_PROFILE_PIC;
-            
+
             if (imageBytes == null)
                 return;
             msisdn = msisdn.Replace(":", "_");
@@ -381,7 +381,7 @@ namespace windows_client.DbUtils
 
             if (msisdn == App.MSISDN)
                 msisdn = HikeConstants.MY_PROFILE_PIC;
-            
+
             msisdn = msisdn.Replace(":", "_");
             using (IsolatedStorageFile store = IsolatedStorageFile.GetUserStoreForApplication())
             {
@@ -404,7 +404,7 @@ namespace windows_client.DbUtils
         {
             if (msisdn == App.MSISDN)
                 msisdn = HikeConstants.MY_PROFILE_PIC;
-            
+
             msisdn = msisdn.Replace(":", "_");
             byte[] data = null;
             using (IsolatedStorageFile store = IsolatedStorageFile.GetUserStoreForApplication())
@@ -710,15 +710,15 @@ namespace windows_client.DbUtils
             {
                 if (!myIsolatedStorage.DirectoryExists(sourceFileDirectory))
                     return;
-                
+
                 if (!myIsolatedStorage.DirectoryExists(destinationFileDirectory))
                     myIsolatedStorage.CreateDirectory(destinationFileDirectory);
-                
+
                 myIsolatedStorage.CopyFile(sourceFilePath, destinationFilePath);
             }
         }
 
-        public static long GetFileSize(string filePath)
+        public static int GetFileSize(string filePath)
         {
             using (IsolatedStorageFile myIsolatedStorage = IsolatedStorageFile.GetUserStoreForApplication())
             {
@@ -726,7 +726,7 @@ namespace windows_client.DbUtils
                 {
                     using (IsolatedStorageFileStream fileStream = myIsolatedStorage.OpenFile(filePath, FileMode.Open, FileAccess.Read))
                     {
-                        return fileStream.Length;
+                        return (int)fileStream.Length;
                     }
                 }
                 else
@@ -769,7 +769,7 @@ namespace windows_client.DbUtils
             string[] attachmentPaths = new string[2];
             attachmentPaths[0] = HikeConstants.FILES_ATTACHMENT + "/" + msisdn;
             attachmentPaths[1] = HikeConstants.FILES_BYTE_LOCATION + "/" + msisdn;
-            
+
             lock (attachmentLock)
             {
                 using (IsolatedStorageFile store = IsolatedStorageFile.GetUserStoreForApplication())
@@ -1185,14 +1185,14 @@ namespace windows_client.DbUtils
                         store.DeleteFile(fName);
                         return;
                     }
-                   
+
                     using (var file = store.OpenFile(fName, FileMode.Create, FileAccess.Write, FileShare.ReadWrite))
                     {
                         using (BinaryWriter writer = new BinaryWriter(file))
                         {
                             writer.Seek(0, SeekOrigin.Begin);
                             writer.Write(App.ViewModel.PicUploadList.Count);
-                            
+
                             foreach (var ms in App.ViewModel.PicUploadList)
                             {
                                 ms.Write(writer);
@@ -1219,9 +1219,9 @@ namespace windows_client.DbUtils
                 {
                     if (!store.DirectoryExists(MISC_DIR))
                         return;
-                    
+
                     string fname = MISC_DIR + "\\" + PENDING_PROFILE_PIC_REQ_FILE;
-                    
+
                     if (!store.FileExists(fname))
                         return;
 
@@ -1230,7 +1230,7 @@ namespace windows_client.DbUtils
                         using (var reader = new BinaryReader(file))
                         {
                             int count = 0;
-                            
+
                             try
                             {
                                 count = reader.ReadInt32();
